+++ cmeel/__main__.py
@@ -9,6 +9,7 @@ import sys
 from . import __version__
 from .docker import add_docker_arguments, docker_cmd
 from .env import add_paths_arguments, paths_cmd
+from .metadata import add_metadata_arguments, metadata_cmd
 from .release import add_release_arguments, release_cmd
 
 LOG = logging.getLogger("cmeel")
@@ -42,6 +43,7 @@ def parse_args() -> argparse.Namespace:
     add_paths_arguments(subparsers)
     add_docker_arguments(subparsers)
     add_release_arguments(subparsers)
+    add_metadata_arguments(subparsers)
 
     ver = subparsers.add_parser("version", help="print current cmeel version.")
     ver.set_defaults(cmd="version")
@@ -71,6 +73,8 @@ def main():
         docker_cmd(**vars(args))
     elif args.cmd == "release":
         release_cmd(**vars(args))
+    elif args.cmd == "metadata":
+        print(metadata_cmd(**vars(args)))
     elif args.cmd == "version":
         print(f"This is cmeel version {__version__}")
     else: