+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - :warning: Refactored python API. No change to cli or pip APIs.
+- added `metadata` cli command
 
 ## [v0.59.0] - 2026-01-19